+++ CHANGELOG.md
@@ -6,7 +6,7 @@ changed functionality, fixed bugs).
 # Version 1.12 (in development)
 
 - Removed "late imports", to help isolate Blender add-ons bundling BAT from each other.
-
+- Decided to *not* support the Shaman API of Flamenco 3.x in BAT. The support for that protocol will be implemented in the Flamenco 3.x add-on for Blender, and not in BAT itself. A new future version of BAT will remove support for the Shaman API altogether.
 
 # Version 1.11 (2022-02-18)